requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
google-generativeai>=0.8.0
openai>=1.50.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
streamlit>=1.38.0
sqlalchemy>=2.0.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
python-dateutil>=2.9.0
tabulate>=0.9.0
rich>=13.0.0
